fix: Refactoring code.
